refactor(excel_to_json): try_parse_number function moved to utils
